refactor: use common hash domain separator
