+++ api/_transfer-with-auth.ts
@@ -1,7 +1,7 @@
 import { BigNumberish, ethers } from "ethers";
 import { getProvider } from "./_utils";
 import { ERC_TRANSFER_WITH_AUTH_ABI } from "./_abis";
-import { utils } from "ethers";
+import { hashDomainSeparator } from "./_eip712";
 
 export class TransferWithAuthNotSupportedError extends Error {
   constructor(tokenAddress: string, cause?: Error) {
@@ -20,28 +20,6 @@ export class TransferWithAuthDomainSeparatorMismatchError extends Error {
       `Transfer with authorization EIP712 domain separator mismatch for ${tokenAddress}`
     );
   }
-}
-
-export function hashDomainSeparator(params: {
-  name: string;
-  version: string | number;
-  chainId: number;
-  verifyingContract: string;
-}): string {
-  return utils.keccak256(
-    utils.defaultAbiCoder.encode(
-      ["bytes32", "bytes32", "bytes32", "uint256", "address"],
-      [
-        utils.id(
-          "EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"
-        ),
-        utils.id(params.name),
-        utils.id(params.version.toString()),
-        params.chainId,
-        params.verifyingContract,
-      ]
-    )
-  );
 }
 
 export async function getReceiveWithAuthTypedData(params: {
